+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.7.0] - 2020-06-05
+
+### Added
+
+- Separate modules for routes and controllers.
+- Server.ts file.
+
+### Changed
+
+- app.ts file.
+- NPM scripts in package.json file to start the application from server.ts file.
+
 ## [0.6.0] - 2020-06-04
 
 ### Added